Tests/LibWeb/Text/expected/context-menu-selects-text.txt

plain word: "ordinary"
url token: "ladybird.org/path?x=1"
canceled context menu: ""
existing selection preserved: "existing"
existing selection replaced: "other"
image preserves selection: "existing"
user-select none: ""
